Auth and other changes

+++ client/src/actions/posts.js
@@ -2,14 +2,33 @@ import { FETCH_ALL, CREATE, UPDATE, DELETE, LIKE } from "../constants/actionType
 import * as api from "../api";
 
 //Action Creators
+
+export const findToken = () => {
+    var token = decodeURIComponent(document.cookie);
+    if(!token || token.length <6 ){
+        return null;
+    }
+
+    token = token.substring(6);
+    return token;
+} 
+
 export const getPosts = () => async(dispatch) => {
 
 	try{
-		const {data} = await api.fetchPosts();
+
+		const token = findToken();
+		if(token === null){
+			return;
+		}
+
+		const {data} = await api.fetchPosts(token);
+
 		dispatch({
 		    type: FETCH_ALL,
 		    payload: data
 	    });
+
 	}catch(error){
 		console.log(error);
 	}
@@ -18,7 +37,13 @@ export const getPosts = () => async(dispatch) => {
 export const createPost = (post) => async(dispatch) => {
 
 	try{
-		const {data} = await api.createPost(post);
+
+		const token = findToken();
+		if(token === null){
+			return;
+		}
+
+		const {data} = await api.createPost(post, token);
 		dispatch({
 		    type: CREATE,
 		    payload: data
@@ -28,10 +53,16 @@ export const createPost = (post) => async(dispatch) => {
 	}
 }
 
-export const updatePost = (id,post) => async(dispatch) => {
+export const updatePost = (id, post) => async(dispatch) => {
 
 	try{
-		const {data} = await api.updatePost(id,post);
+
+		const token = findToken();
+		if(token === null){
+			return;
+		}
+
+		const {data} = await api.updatePost(id, post, token);
 		dispatch({
 		    type: UPDATE,
 		    payload: data
@@ -44,7 +75,13 @@ export const updatePost = (id,post) => async(dispatch) => {
 export const deletePost = (id) => async(dispatch) => {
 
 	try{
-		await api.deletePost(id);
+
+		const token = findToken();
+		if(token === null){
+			return false;
+		}
+
+		await api.deletePost(id, token);
 		dispatch({
 		    type: DELETE,
 		    payload: id
@@ -57,11 +94,29 @@ export const deletePost = (id) => async(dispatch) => {
 export const likePost = (id) => async(dispatch) => {
 
 	try{
-		const {data} = await api.likePost(id);
+
+		const token = findToken();
+		if(token === null){
+			return ;
+		}
+
+		const {data} = await api.likePost(id, token);
 		dispatch({
 		    type: LIKE,
 		    payload: data
 	    });
+	}catch(error){
+		console.log(error);
+	}
+}
+
+export const getSpecificPost = async(id) => {
+	
+	try{
+
+		const token = findToken();
+		const {data} = await api.fetchSpecificPost(id, token);
+		return data;
 	}catch(error){
 		console.log(error);
 	}
